error catching
When the CSV is being written to by C# component, Python will throw a PermissionError "Errno 13" when attempting to read it at the exact same time. Rarely, the fille will not exist during these re-writes, and it will encounter a FileNotFoundError. 

The lines where the CSV is first opened, CSV is reloaded, and when data is being pulled from it may all encounter these errors. These lines have been wrapped in an unending while loop until the exception does not occur. Worst case scenario, the detector just quietly doesn't work if the file actually doesn't exist (which is ideal). 

There's probably a better way to do this.