Fixed loss function
Changed to BCEWithLogitsLoss loss function which allows me to account for weighted data. If you have significantly more "Wins" than "Losses" in your data (or vice versa) you have to tell your model that. This lets me do that with the results column. Summing the entire column gives me the number of wins, number of rows div wins = loss:win rate. 

This needs to be incorporated into the model itself. Print statements need to be removed.